+++ CHANGES.md
@@ -1,6 +1,9 @@
 1.0.2 (In Progress)
 -------------------
 
+#### Changed
+* It is now possible to have the overlay layer be visible over the captured image. This is off by default.
+
 #### Fixed
 * Overlay layer is now correctly shown if it is added to the view before the camera is started.
 * The delegate method for the view weill not be incorrectly called if an image is not captured by the camera when the camera button is pressed.